annotate errors so that we know exactly which line of code failed
It is not necessarily obvious from an error's string representation
what we did to produce that error.

If we always used panic() everywhere, then upon program failure, we would get
a "panic stack" telling us exactly which lines of code failed.

But, it is not idiomatic to use panic() all over the place. We are meant to
return error values instead.

So, it is nice to annotate our error values with file/line information.
It allows the user of our program to pinpoint the failing line of code.